[ecal_service] renaming header files to avoid future name clashes in eCAL core.
