+++ ecal/core/src/service/ecal_service_client_impl.h
@@ -27,7 +27,7 @@
 #include <ecal/v5/ecal_callback.h>
 #include <ecal/ecal_service_info.h>
 #include <ecal/ecal_types.h>
-#include <ecal/service/client_session.h>
+#include <ecal_service/client_session.h>
 
 #include "serialization/ecal_serialize_sample_registration.h"
 #include "serialization/ecal_struct_service.h"
