+++ ecal/service/ecal_service/include/ecal_service/client_session.h
@@ -35,10 +35,10 @@
 #pragma warning(pop)
 #endif
 
-#include <ecal/service/client_session_types.h>
-#include <ecal/service/error.h>
-#include <ecal/service/logger.h>
-#include <ecal/service/state.h>
+#include <ecal_service/client_session_types.h>
+#include <ecal_service/error.h>
+#include <ecal_service/logger.h>
+#include <ecal_service/state.h>
 
 namespace eCAL
 {
