+++ ecal/service/ecal_service/include/ecal_service/client_session_types.h
@@ -23,7 +23,7 @@
 #include <memory>
 #include <string>
 
-#include <ecal/service/error.h>
+#include <ecal_service/error.h>
 
 namespace eCAL
 {
