+++ ecal/service/ecal_service/include/ecal_service/server_session_types.h
