+++ ecal/service/ecal_service/src/client_session_impl_base.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2019 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,9 +32,9 @@
 #pragma warning(pop)
 #endif
 
-#include <ecal/service/client_session_types.h>
+#include <ecal_service/client_session_types.h>
 
-#include <ecal/service/state.h>
+#include <ecal_service/state.h>
 
 namespace eCAL
 {
