+++ ecal/service/ecal_service/src/client_session_impl_v1.cpp
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2024 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -36,10 +36,10 @@
 
 #include <asio.hpp>
 
-#include <ecal/service/client_session_types.h>
-#include <ecal/service/error.h>
-#include <ecal/service/logger.h>
-#include <ecal/service/state.h>
+#include <ecal_service/client_session_types.h>
+#include <ecal_service/error.h>
+#include <ecal_service/logger.h>
+#include <ecal_service/state.h>
 
 #ifdef WIN32
   #include <Winsock2.h>
