+++ ecal/service/ecal_service/src/server.cpp
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2019 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,14 +17,14 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include <ecal/service/server.h>
+#include <ecal_service/server.h>
 
 #include <cstdint>
 #include <memory>
 
 #include <asio.hpp>
 
-#include <ecal/service/logger.h>
+#include <ecal_service/logger.h>
 
 #include "server_impl.h"
