+++ ecal/service/samples/sample_client/src/main.cpp
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2023 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,10 +31,10 @@
 
 #include <asio.hpp>
 
-#include <ecal/service/client_manager.h>
-#include <ecal/service/client_session_types.h>
-#include <ecal/service/error.h>
-#include <ecal/service/logger.h>
+#include <ecal_service/client_manager.h>
+#include <ecal_service/client_session_types.h>
+#include <ecal_service/error.h>
+#include <ecal_service/logger.h>
 
 void print_usage(const std::string& arg0)
 {
