+++ ecal/service/samples/sample_server/src/main.cpp
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2023 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,9 +26,9 @@
 
 #include <asio.hpp>
 
-#include <ecal/service/server_manager.h>
-#include <ecal/service/server_session_types.h>
-#include <ecal/service/logger.h>
+#include <ecal_service/server_manager.h>
+#include <ecal_service/server_session_types.h>
+#include <ecal_service/logger.h>
 
 void print_usage(const std::string& arg0)
 {
