+++ ecal/service/samples/sample_standalone/src/main.cpp
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2023 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,12 +17,12 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include <ecal/service/server_manager.h>
-#include <ecal/service/client_manager.h>
+#include <ecal_service/server_manager.h>
+#include <ecal_service/client_manager.h>
 
-#include <ecal/service/error.h>
-#include <ecal/service/client_session_types.h>
-#include <ecal/service/server_session_types.h>
+#include <ecal_service/error.h>
+#include <ecal_service/client_session_types.h>
+#include <ecal_service/server_session_types.h>
 
 #include <chrono>
 #include <iostream>
